feat: connect user and post

+++ src/main/java/com/project2/board/service/PostService.java
@@ -1,6 +1,8 @@
 package com.project2.board.service;
 
 import com.project2.board.exception.post.PostNotFoundException;
+import com.project2.board.exception.user.UserNotAllowedException;
+import com.project2.board.model.entity.UserEntity;
 import com.project2.board.model.post.Post;
 import com.project2.board.model.post.PostPatchRequestBody;
 import com.project2.board.model.post.PostPostRequestBody;
@@ -33,31 +35,38 @@ public class PostService {
         return Post.from(postEntity);
     }
 
-    public Post createPost(PostPostRequestBody postPostRequestBody) {
-        var postEntity = new PostEntity();
-        postEntity.setBody(postPostRequestBody.body());
-        var savedPostEntity = postEntityRepository.save(postEntity);
-        return Post.from(savedPostEntity);
+    public Post createPost(PostPostRequestBody postPostRequestBody, UserEntity currentUser) {
+        var postEntity = postEntityRepository.save(
+                PostEntity.of(postPostRequestBody.body(), currentUser));
+        return Post.from(postEntity);
     }
 
-    public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody) {
+    public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody, UserEntity currentUser) {
         var postEntity = postEntityRepository
                 .findById(postId)
                 .orElseThrow(
                         () -> new PostNotFoundException(postId)
                 );
+
+        if (!postEntity.getUser().equals(currentUser)) {
+            throw new UserNotAllowedException();
+        }
 
         postEntity.setBody(postPatchRequestBody.body());
         var updatedPostEntity = postEntityRepository.save(postEntity);
         return Post.from(updatedPostEntity);
     }
 
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, UserEntity currentUser) {
         var postEntity = postEntityRepository
                 .findById(postId)
                 .orElseThrow(
                         () -> new PostNotFoundException(postId)
                 );
+
+        if (!postEntity.getUser().equals(currentUser)) {
+            throw new UserNotAllowedException();
+        }
 
         postEntityRepository.delete(postEntity);
     }
